delete publish files
https://stackoverflow.com/a/70668597
I think regarding @Phil's answer, the 2 actions are doing different things.

1. DeleteExistingFiles
This will delete existing files on the web server when you actually deploy. I don't see any documentation about this setting from Microsoft which is frustrating. This same feature is available when you publish your application to a Web Deploy Package and use the Import feature in IIS to import the application. You will get a prompt like this: enter image description here

This is the same as DeleteExistingFiles in a publish profile.

2. Publish directory
Deleting the publish directory is a different thing. The publish directory is a working directory of the msbuild /t:Publish or msbuild /t:Package commands. Basically what is in the publish directory is what ends up in the ZIP file used for a web deploy package or a publish profile. You can get issues during publish process with incorrect files in the publish directory. For example, https://github.com/scottksmith95/LINQKit/issues/157. Unfortunately there is no way to delete or clean the publish directory with the existing msbuild or dotnet cli commands: https://github.com/dotnet/sdk/issues/5972.

Fun times. So @Phil's del command is what we are left with.

Share
Improve this answer
Follow
edited Jan 11, 2022 at 18:13
answered Jan 11, 2022 at 14:37
Jess's user avatar
Jess
23.1k1919 gold badges122122 silver badges141